lambdas/requirements.txt
psycopg2-binary==2.9.10
requests==2.31.0
pydantic==2.6.1
